190 ams to bms, stack sizes power of 2
