+++ Core/Inc/cerberus_conf.h
@@ -16,7 +16,7 @@
 #define SERIAL_MONITOR_DELAY
 #define CAN_ROUTER_DELAY
 #define CAN_DISPATCH_DELAY    5
-#define AMS_CAN_MONITOR_DELAY 4000
+#define BMS_CAN_MONITOR_DELAY 4000
 #define STATE_MACHINE_DELAY
 #define TORQUE_CALC_DELAY
 #define FAULT_HANDLE_DELAY
