+++ Core/Inc/fault.h
@@ -19,7 +19,7 @@ typedef enum {
 	STEERINGIO_ROUTING_FAULT = 0x100,
 	STATE_RECEIVED_FAULT = 0x200,
 	INVALID_TRANSITION_FAULT = 0x400,
-	AMS_CAN_MONITOR_FAULT = 0x800,
+	BMS_CAN_MONITOR_FAULT = 0x800,
 	BUTTONS_MONITOR_FAULT = 0xF00,
 	BSPD_PREFAULT = 0x1000,
 	LV_MONITOR_FAULT = 0x2000,
