+++ Makefile
@@ -52,7 +52,7 @@ Core/Src/can_handler.c \
 Core/Src/serial_monitor.c \
 Core/Src/dti.c \
 Core/Src/state_machine.c \
-Core/Src/ams.c \
+Core/Src/bms.c \
 Core/Src/nero.c \
 Core/Src/pdu.c \
 Core/Src/mpu.c \
